Cuatro cosas que señaló la revisión, y partir el auditor en dos
El auditor cruzaba las 800 líneas y el trinquete no lo permite. Recortar
comentarios para colar el número sería trampa, así que se parte por donde
tiene sentido: la auditoría de líneas fiscales mira otro store y contesta otra
pregunta, y se va a `__lineasFiscalesAudit`. Los comparadores compartidos
salen a un módulo aparte para que los dos no acaben con criterios distintos —
que es la clase de fallo que estas herramientas existen para encontrar.

De la revisión:

  - Dos cargos del mismo mes compartían `origenId` y el segundo pisaba al
    primero, dejando el mes valiendo el último en vez de los dos. Hoy ningún
    patrón devuelve dos fechas en un mes, pero la clave sí lo permite: se
    agrupa por mes y se SUMA, que es lo que ya hace `importeCompromisoEnMes`.
  - Las líneas antiguas sin `origenId` se reconocían por el alias del gasto.
    Al renombrarlo dejaban de casar y se creaba una línea nueva encima de la
    vieja, duplicando el mes. Ahora se ADOPTAN —se les escribe el `origenId`—
    mirando sólo casilla y mes, que no cambian al renombrar. Con más de una
    candidata no se toca ninguna: dos recurrentes pueden compartir casilla y
    mes, y adoptar a ciegas reescribiría la del otro.
  - La migración escribía fila a fila; pasa a una transacción única, que corre
    en el arranque y además no deja la tabla medio corregida si falla.
  - `localStorage` se leía fuera del `try`: en un iframe o en modo privado
    puede lanzar y tumbar el arranque antes del `catch`.

Y la cabecera del auditor decía «NO importar desde código de producción»
mientras la página de Tesorería lo importa. Lo que no debe hacerse es
llamarlo desde lógica de negocio; registrarlo en la consola de producción es
justamente su razón de ser, porque ahí es donde están los datos.
